pyproject.toml
[project]
name = "appdownloader"
version = "0.1.0"
description = "Internal APK upload/download portal for company networks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.129.2",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib[argon2]>=1.7.4",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
appdownloader = "appdownloader:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["uv_build>=0.9.12,<0.10.0"]
build-backend = "uv_build"
